12 - error component 作成
問題
複数の箇所で同じコードが繰り返し使用されていた

解決策
エラーメッセージを表示するcomponentを作成した
