+++ src/components/common/Error.js
@@ -1,0 +1,19 @@
+import React from 'react';
+import { Text } from 'react-native';
+
+const Error = ({ message }) => (
+  <Text style={styles.errorTextStyle}>
+    {message}
+  </Text>
+);
+
+const styles = {
+  errorTextStyle: {
+    paddingTop: 5,
+    fontSize: 18,
+    alignSelf: 'center',
+    color: '#FF0000',
+  }
+};
+
+export { Error };
