+++ src/components/common/index.js
@@ -5,3 +5,4 @@ export * from './Header';
 export * from './Input';
 export * from './Spinner';
 export * from './Confirm';
+export * from './Error';
